Add client maintenance UI and helpers
Introduce a set of maintenance UI components and shared helpers for the client app. Adds CreateMaintenanceModal, ClientDetailModal, and MaintenanceBadges (Toast, Loader, Empty, PBadge, SBadge, Pager, Field, Row2, Modal) to handle request creation, detail viewing and common UI bits. Adds maintenanceHelpers with categories/priorities, styling constants, global CSS string and formatting utilities (fmtMoney, fmtDT, fmtRel). Refactors src/pages/client/Maintenance.jsx to import and use the new components/utilities instead of inline definitions. CreateMaintenanceModal posts to /api/maintenance and uses notifications on validation/errors.
